- player classes derived from DoomPlayer now are affected by buckshot and instagib - player classes derived from DoomPlayer or HereticPlayer get the LMS weapons in LMS games - fixed a few warnings
SVN r891 (trunk)

+++ src/network.cpp
@@ -555,7 +555,7 @@ int NETWORK_GetPackets( void )
 		SERVER_STATISTIC_AddToInboundDataTransfer( lNumBytes );
 
 	// If the number of bytes we're receiving exceeds our buffer size, ignore the packet.
-	if ( lNumBytes >= g_NetworkMessage.ulMaxSize )
+	if ( lNumBytes >= static_cast<LONG>(g_NetworkMessage.ulMaxSize) )
 		return ( 0 );
 
 	// Decode the huffman-encoded message we received.
@@ -629,7 +629,7 @@ int NETWORK_GetLANPackets( void )
 		SERVER_STATISTIC_AddToInboundDataTransfer( lNumBytes );
 
 	// If the number of bytes we're receiving exceeds our buffer size, ignore the packet.
-	if ( lNumBytes >= g_NetworkMessage.ulMaxSize )
+	if ( lNumBytes >= static_cast<LONG>(g_NetworkMessage.ulMaxSize) )
 		return ( 0 );
 
 	// Decode the huffman-encoded message we received.
